+++ .gitignore
@@ -1,7 +1,5 @@
 .idea
-*.node
-sandbox
+.coverage
 node_modules
-coverage
 test/**/*.generated.*
 test/**/*.min.css
