+++ .travis.yml
@@ -5,6 +5,4 @@ node_js:
   - "8"
   - "7"
   - "6"
-  - "5"
-  - "4"
-after_script: "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
+after_script: "cat ./.coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
