+++ package.json
@@ -2,7 +2,7 @@
   "name": "gulp-clean-css",
   "description": "Minify css with clean-css.",
   "homepage": "https://github.com/scniro/gulp-clean-css#readme",
-  "version": "3.10.0",
+  "version": "4.0.0",
   "author": "scniro",
   "license": "MIT",
   "bugs": {
@@ -28,31 +28,41 @@
     "gulpplugin",
     "gulpfriendly"
   ],
+  "nyc": {
+    "report-dir": "./.coverage",
+    "reporter": [
+      "lcov",
+      "text"
+    ],
+    "include": [
+      "index.js"
+    ],
+    "sourceMap": false,
+    "temp-dir": "./.coverage/.nyc_output"
+  },
   "dependencies": {
     "clean-css": "4.2.1",
     "plugin-error": "1.0.1",
-    "through2": "2.0.3",
+    "through2": "3.0.0",
     "vinyl-sourcemaps-apply": "0.2.1"
   },
   "devDependencies": {
-    "chai": "4.1.2",
-    "chai-string": "1.4.0",
+    "chai": "4.2.0",
+    "chai-string": "1.5.0",
     "coveralls": "3.0.2",
-    "express": "4.16.3",
+    "express": "4.16.4",
     "fancy-log": "1.3.2",
-    "gulp": "3.9.1",
+    "gulp": "4.0.0",
     "gulp-concat": "2.6.1",
-    "gulp-istanbul": "1.1.3",
-    "gulp-mocha": "6.0.0",
     "gulp-rename": "1.4.0",
-    "gulp-sass": "4.0.1",
+    "gulp-sass": "4.0.2",
     "gulp-sourcemaps": "2.6.4",
     "mocha": "5.2.0",
-    "vinyl": "2.2.0",
-    "vinyl-buffer": "1.0.1",
-    "vinyl-fs-fake": "1.1.0"
+    "nyc": "13.1.0",
+    "vinyl": "2.2.0"
   },
   "scripts": {
-    "test": "./node_modules/.bin/gulp test"
+    "cover": "./node_modules/.bin/nyc npm test",
+    "test": "./node_modules/.bin/mocha ./index.spec.js"
   }
 }
