+++ test/fixtures/malformed.css
@@ -1,0 +1,1 @@
+@import url(/some/fake/file);
